cleaned up and renamed trustchainactivity
