+++ app/src/main/AndroidManifest.xml
@@ -21,7 +21,7 @@
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity
-            android:name=".main.TrustChainActivity"
+            android:name=".main.PeerOverviewActivity"
             android:configChanges="orientation|screenLayout|screenSize"
             android:windowSoftInputMode="stateUnchanged" />
         <activity android:name=".chainExplorer.ChainExplorerActivity" />
