+++ app/src/main/java/nl/tudelft/cs4160/trustchain_android/Network/NetworkCommunicationListener.java
@@ -1,7 +1,6 @@
 package nl.tudelft.cs4160.trustchain_android.Network;
 
 import java.io.IOException;
-import java.net.InetSocketAddress;
 
 import nl.tudelft.cs4160.trustchain_android.appToApp.PeerAppToApp;
 import nl.tudelft.cs4160.trustchain_android.appToApp.PeerHandler;
